requirements.txt
basemap==1.4.1
basemap_data==1.3.2
boto3==1.26.146
botocore==1.29.146
cdsapi==0.6.1
cfgrib==0.9.10.4
dict_recursive_update==1.0.1
eccodes==1.5.1
einops==0.8.0
imageio==2.9.0
loguru==0.7.2
matplotlib==3.7.1
MetPy==1.6.1
mmcv==2.1.0
mmengine
netCDF4==1.6.3
numpy==1.25.0rc1
onnxconverter_common==1.14.0
onnxruntime_gpu==1.16.3
onnxsim==0.4.36
pandas==2.0.3
psutil==5.9.5
pynvml==11.5.0
pytz==2023.3
scipy==1.14.0
timm==0.9.2
tqdm==4.66.4
tvm==1.0.0
urllib3==1.26.16
xarray==2023.7.0
ecmwflibs==0.5.3
